requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.8
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
slowapi==0.1.9
orjson==3.10.7
loguru==0.7.2
openai>=1.40.0
fakeredis==2.23.2
pytest==8.2.1
